Big reformats
I've swapped out lists for headings and sub headings.

They format better and can be linked to.

I've converted `inline code` to:

```
code blocks
```

because they render with a little button to copy the contents to
clipboard

Where we're instructing the student to append text to a file (mostly
~/.zshrc), I've used `cat << EOF >> ~/.zshrc` instead of VS Code.

This sacrifices a little bit of practice with VS Code, but I'm
prioritising getting the setup done consistently and quickly here.